move reader into dataset
