+++ tests/clickhouse/test_query.py
@@ -1,7 +1,7 @@
 from tests.base import BaseEventsTest
 from unittest.mock import patch
 
-from snuba.clickhouse.query import DictClickhouseQuery
+from snuba.clickhouse.dictquery import DictClickhouseQuery
 from snuba.query.query import Query
 from snuba.request.request_settings import HTTPRequestSettings
