+++ tests/query/processors/test_tags_processor.py
@@ -1,7 +1,7 @@
 import pytest
 
 from snuba import state
-from snuba.clickhouse.query import DictClickhouseQuery
+from snuba.clickhouse.dictquery import DictClickhouseQuery
 from snuba.datasets.factory import get_dataset
 from snuba.query.parser import parse_query
 from snuba.request.request_settings import HTTPRequestSettings
